// src/contract/type.ts
export interface IPrice {
  symbol: string;
  amount: number;
}
// ...
export enum ContractMethodType {
  SEND,
  VIEW,
}

export interface IContractOptions {
  chain?: Chain | null;
  type?: ContractMethodType;
}

export interface IGetAllowanceResponse {
  symbol: string;
  owner: string;
  spender: string;
  allowance: number;
}

export interface IContractError extends Error {
  code?: number;
  error?:
    | number
    | string
    | {
        message?: string;
      };
  errorMessage?: {
    message: string;
    name?: string;
    stack?: string;
  };
  Error?: string;
  from?: string;
  sid?: string;
  result?: {
    TransactionId: string;
    transactionId?: string;
  };
  TransactionId?: string;
  transactionId?: string;
  value?: any;
}
// ...
export interface ITimestamp {
  seconds: number;
  nanos: number;
}

export interface IListDuration {
  startTime: ITimestamp;
  publicTime: ITimestamp;
  durationHours?: number;
  durationMinutes: number;
}

export interface IListWithFixedPriceParams {
  symbol: string;
  price: IPrice;
  quantity?: number;
  duration: IListDuration;
  whitelists: {
    whitelists: {
      priceTag: {
        tagName: string;
        price: {
          amount: BigNumber;
          symbol?: string | undefined;
        };
      };
      addressList: {
        value: string[] | undefined;
      };
    }[];
  } | null;
  isWhitelistAvailable: boolean;
}

export interface IGetTotalOfferAmountParams {
  address: string;
  priceSymbol: string;
}

export interface IGetTotalEffectiveListedNFTAmountParams {
  symbol: string;
  address: string;
}

export enum ListType {
  NOT_LISTED,
  FIXED_PRICE,
  ENGLISH_AUCTION,
  DUTCH_AUCTION,
}

export interface IListedNFTInfo {
  symbol: string;
  owner: string;
  quantity: number;
  listType: ListType;
  price: IPrice;
  duration: IListDuration;
}
// ...
export interface ITransactionLog {
  Address: string;
  Name: string;
  Indexed: string[];
  NonIndexed: string;
}

export interface ITransactionResult {
  TransactionId: string;
  Status: string;
  Logs: ITransactionLog[];
  [props: string]: any;
}

export interface ISendResult {
  TransactionId: string;
  TransactionResult: ITransactionResult;
}

export interface IGetProxyAccountByProxyAccountAddressRes {
  createChainId: number;
  managementAddresses: {
    address: string;
  }[];
  proxyAccountHash: string;
}

export interface IForwardCallParams {
  proxyAccountHash: string;
  contractAddress: string;
  methodName: string;
  args: any;
}

export interface IAuctionParams {
  auctionId: string;
  amount: number;
}

export interface IFixPriceList {
  offerTo: string;
  quantity: number;
  price: IPrice;
  startTime: ITimestamp;
}

export interface IBatchBuyNowParams {
  symbol: string;
  fixPriceList: IFixPriceList[];
}